Am rezolvat toate comentariile

+++ Tema1/src/inventar.cpp
@@ -20,7 +20,7 @@ void inventar::CalculInv()
     std::vector<undita>v2 = this->und;
     std::vector<carlig>v3 = this->carr;
     std::vector<mamaliga>v4 = this->mam;
-    for(int i=0; i<v2.size(); i++)
+    for(int i=0; i<(int)v2.size(); i++)
     {
         x = x + v2[i].getStoc()*v2[i].getPret() + v3[i].getStoc()*v3[i].getPret() + v4[i].getStoc()*v4[i].getPret();
     }
@@ -34,7 +34,7 @@ void inventar::CreareMonturaIdeala()
     std::vector<mamaliga>v4 = this->mam;
     std::cout<<"Montura ideala : " <<std::endl<<"================"<<std::endl<<std::endl;
     int maximund=0,maximcarr=0,maximmam=0,indiceund,indicecarr,indicemam;
-    for(int i=0; i<v2.size(); i++)
+    for(int i=0; i<(int)v2.size(); i++)
     {
         if(v2[i].getPret()>maximund)
         {
